+++ lib/index.ts
@@ -1,6 +1,6 @@
 import { GaussianBlurBackgroundProcessor, GaussianBlurBackgroundProcessorOptions } from './processors/background/GaussianBlurBackgroundProcessor';
 import { VirtualBackgroundProcessor, VirtualBackgroundProcessorOptions } from './processors/background/VirtualBackgroundProcessor';
-import { ImageFit } from './types';
+import { HysteresisConfig, ImageFit } from './types';
 import { isSupported } from './utils/support';
 import { version } from './utils/version';
 
@@ -19,6 +19,7 @@ if (typeof window !== 'undefined') {
 export {
   GaussianBlurBackgroundProcessor,
   GaussianBlurBackgroundProcessorOptions,
+  HysteresisConfig,
   ImageFit,
   isSupported,
   version,